CHG copy: provide
copy checks if provide is necessary by evaluating ::provide_if. If that
prints 0 to stdout, then we need to provide .

The provide_exe can now fail gracefully: If exit value is 1, we assume
that providing  failed with user consent, if it is 5, we assume that it
failed with user ignoral, else it still fails badly.

This behaviour is what we get if provide_exe returns the return value of
zenity: 0 = OK, 1 = cancel, 5 = timeout; everything else should be an
actual error. (The rational for zenity is that it should tell you to go
somewhere where YOU can provide the target or source.)